Moved register to utils
