--- pyActigraphy/register.py
@@ -1,11 +1,0 @@
-from collections import defaultdict
-
-REGISTER = defaultdict(list)
-
-def register(name):
-    def wrapper(f):
-        print('-> Function {} registered in {}'
-              .format(f, name))
-        REGISTER[name].append(f)
-        return f
-    return wrapper
